TrieTest.java
import static org.junit.Assert.assertArrayEquals;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

import org.junit.Before;
import org.junit.Test;

/**
 * @author William Zhuang
 */
public class TrieTest {
    /**
     * Tests that Trie works properly.
     */
    @Test
    public void testTrie() {
        Trie t = new Trie();
        t.insert("hello");
        t.insert("hey");
        t.insert("goodbye");
        System.out.println(t.find("hell", false));
        System.out.println(t.find("hello", true));
        System.out.println(t.find("good", false));
        System.out.println(t.find("bye", false));
        System.out.println(t.find("heyy", false));
        System.out.println(t.find("hell", true));   

    }

    public static void main(String[] args) {
        jh61b.junit.textui.runClasses(TrieTest.class);
    }
}
